transform-agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-bank-transform-agent"
version = "1.0.0"
description = "Data format conversion and validation — x402-paid, MCP-native"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]",
    "orjson",
    "pydantic>=2.0",
    "polars",
    "lxml",
    "ruamel.yaml",
    "tomli-w",
    "pymupdf",
    "openpyxl",
    "python-docx",
    "markdown-it-py",
    "beautifulsoup4",
    "jsonschema",
    "x402",
    "redis[asyncio]",
    "httpx",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "ruff",
]

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
